backend/requirements.txt
backcall==0.1.0
decorator==4.3.2
Django==2.1.7
django-cors-headers==2.5.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
gunicorn==19.9.0
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
